2.0.2, 2.0.4: clarify -d:useMalloc requirement
The 2.0.2 release post noted a particular problem, saying it'd be fixed
in 2.0.4. It wasn't, so reduce the potential confusion: amend the note
for 2.0.2, and add the same note for 2.0.4 (on the basis that if it was
worth mentioning for 2.0.2, I guess it's still worth mentioning later).

My understanding is that the regression still exists in 2.0.4, and Araq
said on the forum [1]:

    This example is supposed to work and does when you disable Nim's
    broken default allocator.

    That is a regression with the 2.0 line. You really need -d:useMalloc
    for the time being, sorry. I've "fixed" this bug months ago but it
    causes other test regressions.

See also the commits between 2.0.2 and 2.0.4 [2].

[1] https://forum.nim-lang.org/t/11374#74050
[2] https://github.com/nim-lang/Nim/compare/v2.0.2...v2.0.4